separate fqdn-list
