+++ catalystwan/models/policy/__init__.py
@@ -14,6 +14,7 @@ from catalystwan.models.policy.list.color import ColorList
 from catalystwan.models.policy.list.communities import CommunityList, ExpandedCommunityList
 from catalystwan.models.policy.list.data_ipv6_prefix import DataIPv6PrefixList
 from catalystwan.models.policy.list.data_prefix import DataPrefixList
+from catalystwan.models.policy.list.fqdn import FQDNList
 
 from .centralized import CentralizedPolicy, TrafficDataDirection
 from .definition.access_control_list import AclPolicy
@@ -31,7 +32,6 @@ from .definition.traffic_data import TrafficDataPolicy
 from .definition.vpn_membership import VPNMembershipPolicy
 from .definition.zone_based_firewall import ZoneBasedFWPolicy
 from .lists import (
-    FQDNList,
     GeoLocationList,
     IPSSignatureList,
     IPv6PrefixList,
